requirements.txt
packaging
torch==1.5.0
transformers==4.3.3
datasets==1.2.1
nltk==3.5
absl-py==0.12.0
rouge_score==0.0.4
sacrebleu==1.5.1
dgl-cu101==0.6.1